% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitPPCurve.R
\name{fitPPCurve}
\alias{fitPPCurve}
\title{Modelfree derivate to fit a psychometric curve}
\usage{
fitPPCurve(stim_level, nb_corr, nb_trials)
}
\arguments{
\item{stim_level}{A vector of the different level of the physical
stimulation (e.g. luminance, sound intensity...). If strings are
used, the function will convert the stings as simple digits.}

\item{nb_corr}{A vector of the number of 'correct' responses of a
given subject in a given condition.}

\item{nb_trials}{A vector of number of trials of a given subject
in a given condition (should be of the same length than the nb_corr).}
}
\value{
Return a list with a vector of fitted value for each point
 of the initial nb_corr variable and the extracted slope, PSS and JND values.
}
\description{
This function use the 'modelfree'package to fit locally
the data of a psychophysic experiment
(see \url{http://personalpages.manchester.ac.uk/staff/d.h.foster/software-modelfree/latest/index.html)}
}
\examples{
### Generate fake date to use the function ---------------------------------
# Levels of the physical stimulation
stim = 1:10
# Number of correct responses for each level of stim
corr = c(0, 0, 3, 5, 10, 14, 16, 18, 19, 20)
# Number of trial per condition actually included
trials = c(20, 20, 19, 20, 18, 19, 20, 20, 19, 20)

### Fitting the curve with the modelfree adapation -------------------------
fitted = fitPPCurve(stim, corr, trials)
fitted

### Visuallizing the fit ---------------------------------------------------
plot(stim, corr/trials)
lines(stim, fitted$pfit, col='red')
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, Universit\'{e} de Montr\'{e}al

Maintainer: Guillaume T. Vallet \email{gtvallet@gmail.com}
}
\seealso{
See \code{vignette("locglmfit", package = "modelfree")} for more details on
the fit function used.
}
\keyword{psychophysic}

