% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotPPCurve}
\alias{plotPPCurve}
\title{A custom ggplot2 wrapper}
\usage{
plotPPCurve(dt, vars = NULL, xvar = NULL, resp = "resp", se = F,
  axnames = NULL)
}
\arguments{
\item{dt}{A data frame with mean already computed}

\item{vars}{A vector of variables to use including the subjects}

\item{xvar}{A variable to designate the x axis}

\item{resp}{A string of the column name the dependent variable (VD) (y axis)}

\item{se}{A logical value to display standard error (should be called se) Defaults to False}

\item{axnames}{A vector of length 2 with the stings to use as y label and x label}
}
\description{
This function allows you to create a ggplot psychometric curve
 with automatic theming and addition of lines of the objective
 point of equivalance. Optionnaly, the standard errors can be added.
}
\examples{
plotme()
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}
}
\keyword{ggplot2}

