% Generated by roxygen2 (4.0.0): do not edit by hand
\name{psychophy}
\alias{psychophy}
\title{Psychophy processes psychophysic raw data to fitted value}
\usage{
psychophy(data, wid = "subject_nr", stim = NULL, resp = "correct",
  vars = NULL, axnames = NULL)
}
\arguments{
\item{data}{A data frame in long format (one row per record).}

\item{wid}{A string indicating the column name of the subjects
or observations. Defaults to "subject_nr".}

\item{stim}{A string indicating the column name of the physical
stimulation (e.g. luminance, sound intensity...).}

\item{resp}{A string indicating the column name of the response
as 0 and 1. The function will remove all values below 0 and above 1.
Defaults to "correct".}

\item{vars}{A vector of string indicating the column name of the
dependent variables.}

\item{axnames}{A vector of length 2 with the stings to use as
y label and x label in the graphs. The function automatically add
"Proportion of " for the y label. Defaults to NULL.}
}
\value{
Return a list with a data frame of raw data, a data frame
 of descriptive data by subject and condition, a data frame of fitted
 values (slopes and PSS) and a list of two graphics, by subjects and global.
}
\description{
This function processes psychophysic raw data to compute
 descriptive data by condition and by subjects; fit the data
 by subject; and extract the slopes and PSS index.
}
\details{
The fitting is done using the fitPPCurve function of the
 present package which use the 'modelfree' package to fit
 the data locally
(see \url{http://personalpages.manchester.ac.uk/staff/d.h.foster/software-modelfree/latest/index.html)}
}
\examples{
### Generate fake date to use the function ---------------------------------
# Levels of the physical stimulation
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}

Maintainer: Guillaume T. Vallet \email{gtvallet@gmail.com}
}
\keyword{psychophysic}

